Split sphinxlint.py file.
